clippy: fix warnings across the codebase
- Use div_ceil() instead of manual (x + d - 1) / d pattern
- Remove unnecessary u64/u32 casts where types already match
- Use writeln!() instead of write!() with trailing \n
- Collapse nested if statements into && chains
- Derive Default for UnlockPolicy instead of manual impl
- Convert while-let-on-iterator to for loop in strip_alloc
- Use rev().find() instead of filter().last() in timestats
- Inline trailing literal args in println! format strings
- Simplify redundant field names in struct initialization

Reduces clippy warnings from 28 to 6 (remaining are deliberate FFI
patterns and the build script).

+++ src/wrappers/handle.rs
@@ -161,9 +161,9 @@ impl BcachefsHandle {
         // Drop path_fd — we'll re-open via sysfs/ctl
         drop(path_fd);
 
-        let mode = stat.st_mode & libc::S_IFMT as u32;
+        let mode = stat.st_mode & libc::S_IFMT;
 
-        if mode == libc::S_IFBLK as u32 {
+        if mode == libc::S_IFBLK {
             // Block device: try sysfs symlink
             let major = rustix::fs::major(stat.st_rdev);
             let minor = rustix::fs::minor(stat.st_rdev);
